+++ front/app/modules/commercial/user_custom_fields/admin/components/AreaChart.tsx
