+++ front/app/modules/commercial/user_custom_fields/admin/components/GenderChart.tsx
@@ -9,13 +9,13 @@ import shallowCompare from 'utils/shallowCompare';
 // intl
 import { FormattedMessage, injectIntl } from 'utils/cl-intl';
 import { InjectedIntlProps } from 'react-intl';
-import messages from '../../messages';
+import messages from '../../../../../containers/Admin/dashboard/messages';
 
 // styling
 import { withTheme } from 'styled-components';
 
 // components
-import ExportMenu from '../../components/ExportMenu';
+import ExportMenu from '../../../../../containers/Admin/dashboard/components/ExportMenu';
 import {
   NoDataContainer,
   GraphCardHeader,
@@ -23,7 +23,7 @@ import {
   GraphCard,
   GraphCardInner,
   PieChartStyleFixesDiv,
-} from '../..';
+} from '../../../../../containers/Admin/dashboard';
 import { PieChart, Pie, Tooltip, Cell, ResponsiveContainer } from 'recharts';
 
 // services
